+++ src/core/shared/locale/languages/en.ts
@@ -128,5 +128,19 @@ export default {
         LESS_THAN_CASH: `Value must be less than cash`,
         LESS_THAN_BANK: `Value must be less than bank`,
         USER_ID_POSITIVE: `User ID must be positive`
+    },
+    [LOCALE_KEYS.WEBVIEW_CHARACTERS]: {
+        LABEL_DELETE: 'Delete n Stuff',
+        LABEL_NEW: 'New',
+        LABEL_SELECT: 'Select',
+        LABEL_YES: 'Yes',
+        LABEL_NO: 'No',
+        LABEL_CONFIRM_DELETE: 'Are you sure you want to delete your character ', // Leave space at end.
+        LABEL_NAME: 'Name',
+        LABEL_AGE: 'Age',
+        LABEL_GENDER: 'Gender',
+        LABEL_HOURS: 'Hours',
+        LABEL_CASH: 'Cash',
+        LABEL_BANK: 'Bank'
     }
 };