+++ src/core/shared/locale/languages/keys.ts
@@ -104,5 +104,6 @@ export const LOCALE_KEYS = {
     // =====================
     // WebView Locales
     // =====================
-    WEBVIEW_ATM: 'webview-atm'
+    WEBVIEW_ATM: 'webview-atm',
+    WEBVIEW_CHARACTERS: 'webview-characters'
 };